Refactor some common code to a separate file.
